Fixed memory leak in MarkovSolver unit test. Turned out to be a simple issue of freeing memory for the exponential after each call to the compute function.
Also managed to further optimize the matrix inverse function. There is now a
special function for computing the product of an upper triangular matrix with a
lower triangular matrix. The structure of this multiplication is such that
the product can be computed in place, without the need to allocate memory for a
new matrix. The result is a further reduction in call time to matInv from 2.95
us to 2.19 us because of fewer memory allocations. 
The new matrix multiplication in place of the old one now takes 8 us per call 
instead of the older 80 us.