Addition of the MarkovSolverBase class. The separation of the matrix exponential calculation from the rest of the housekeeping dealing with the exponential lookup tables is complete.
Any Markov solver class can derive from MarkovSolverBase, but has to implement
the computeMatrixExponential() method, which will calculate the exponential of
the instantaneous rate matrix Q. 

Couple more changes to the Interpol2D class, which were not documented in the
previous log entry : 
The getTableValue function now wraps around when the indices given to it are
out of range i.e. any index greater than the size of the table will return the
last element in that dimension. This was necessitated by the MarkovSolverBase
class' need to access all elements of the 2D lookup table while bypassing the 
MOOSE interface of the Interpol2D class.