Some design change in the multiscaleGUI that I had overlooked before:
For MDI, instead of having multiple editor widgets under a single
plugin, we should have multiple instances of a plugin each with just
one editor widget, one plot widget and one runwidget. 

Now we associate a plugin instance with an element of the moose model
and the plugin handles display/editing of the model tree rooted at the
root.

If we need to edit another model using the same plugin class, we
create another instance.

Major changes:

1. MoosePluginBase
  a) constructor takes the keyword arguments:
root: string path for the model root

  b) setCurrentView takes a string name for the view.  This is so that
     plugin developers can implement more views than we can
     foresee. They can provide appropriate menu entries for each
     possible view without going though the main window.

2. ViewBase 

   a) now takes a plugin as parameter and stores it in the "plugin"
   field for future reference.

   b) getCentralWidgets changed to getCentralWidget as each view will
   provide only one central widget.


3. EditorWidgetBase - update() renamed to updateModelView() to avoid
name collision with QWidget's method.

4. MoosePlugin - updated to correspond to above.