Add "hsolveCuda" folder and change Makefile to compile this folder with USE_CUDA=1 command line flag.
The hsolveCuda folder contains:

cudaLibrary, a separate project to make a shared library. Because this part of code should compile with NVCC compiler, using this mechanism available us to use CUDA without changing in normal compile way.

Codes which should execute in CPU part and also API of shared library

Test, a test project to load a Python test model and execute it. Python debugger in Eclipse is not compatible with C++ future breakpoints. Therefore, by directly debugging the test project in Eclipse we can follow the control flow inside moose.